+++ src/components/episode/scenes/vetoCeremonyScene.tsx
@@ -1,5 +1,5 @@
 import { GameState, Houseguest, getById, exclude } from "../../../model";
-import { Scene } from "../scene";
+import { Scene } from "./scene";
 import { useGoldenVeto, backdoorNPlayers } from "../../../utils/ai/aiApi";
 import { Portrait } from "../../playerPortrait/portraits";
 import { NextEpisodeButton } from "../../nextEpisodeButton/nextEpisodeButton";
